requirements.txt
fastapi
uvicorn
smbus2
RPi.bme280
